requirements.txt
pii-data >= 0.5.0, <1.0.0

pii-preprocess >= 0.1.0, <1.0.0
pii-extract-base >= 0.7.0, <1.0.0
pii-extract-plg-regex >= 0.5.1, <1.0.0
pii-decide >= 0.1.0, <1.0.0
pii-transform >= 0.7.0, <1.0.0

transformers: pii-extract-plg-transformers >= 0.1.3, <1.0.0
presidio: pii-extract-plg-presidio >= 0.3.3, <1.0.0
